% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\alias{is_plumber_api}
\title{Create a new plumber API, optionally based on one or more plumber files}
\usage{
api(
  ...,
  host = get_opts("host", "127.0.0.1"),
  port = get_opts("port", 8080),
  env = caller_env()
)

is_plumber_api(x)
}
\arguments{
\item{...}{plumber files or directories containing plumber files to be parsed
in the given order. The order of parsing determines the final order of the
routes in the stack}

\item{host}{A string that is a valid IPv4 address that is owned by this
server}

\item{port}{A number or integer that indicates the server port that should be
listened on. Note that on most Unix-like systems including Linux and macOS,
port numbers smaller than 1024 require root privileges.}

\item{env}{The environment the files should be parsed in}

\item{x}{An object to test for whether it is a plumber api}
}
\value{
A Plumber object
}
\description{
This is the main way to create a new Plumber object that encapsulates your
full api.
}
