% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_auth.R
\name{api_authenticator}
\alias{api_authenticator}
\title{Add an authentication mechanism to your API}
\usage{
api_authenticator(api, auth, name = NULL)
}
\arguments{
\item{api}{A plumber2 api object to add the authenticator to}

\item{auth}{An \link[fireproof:Auth]{Auth} subclass object defining an
authentication scheme}

\item{name}{The name to use for referencing the scheme in an authentication
flow}
}
\value{
This functions return the \code{api} object allowing for easy chaining
with the pipe
}
\description{
This function adds an authentication scheme to your API. Notably, this does
not turn on authentication for any of your handlers but makes it available
for reference in an authentication flow. To use it, reference it in the
\code{auth_flow} argument of functions supporting it. Authentication schemes are
defined using the various \verb{auth_*()} constructors in the fireproof package.
Refer to these for further documentation
}
\section{Using annotation}{
To add an authenticator to your api defined in an annotated file use the
\verb{@authenticator} tag:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#* @authenticator BasicAuth
fireproof::auth_basic(...)
}\if{html}{\out{</div>}}

The tag parameter (\code{BasicAuth}) provides the name for the authenticator
}

\examples{
auth <- fireproof::auth_key("plumber2-key", "MY_VERY_SECRET_KEY")

api() |>
  api_authenticator(auth, "cookie_key")

}
