% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{async_evaluaters}
\alias{async_evaluaters}
\alias{future_async}
\alias{mirai_async}
\title{Async evaluaters provided by plumber}
\usage{
future_async(substitute = FALSE, ...)

mirai_async(...)
}
\arguments{
\item{...}{Further argument passed on to the internal async function.
See Details for information on which function handles the formatting
internally in each async evaluater}
}
\value{
A function taking \code{expr} and \code{envir}. The former is the expression to
evaluate and the latter is an environment with additional variables that
should be made available during evaluation
}
\description{
These functions support async request handling. You can register your own as
well using \code{\link[=register_async]{register_async()}}.
}
\section{Provided evaluaters}{
\itemize{
\item \code{future_async()} uses \code{\link[promises:future_promise]{promises::future_promise()}}. It is registered as
\code{"future"}. Be aware that for this evaluater to execute asynchronously you
need to set a different planner than the default. See \code{\link[future:plan]{future::plan()}}.
\item \code{mirai_async()} uses \code{\link[mirai:mirai]{mirai::mirai()}}. It is registered as
\code{"mirai"}. Be aware that for this evaluater to be performant you should
start up multiple persistent background processes. See \code{\link[mirai:daemons]{mirai::daemons()}}.
}
}

