% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers.R
\name{init_formatter}
\alias{init_formatter}
\alias{close_formatter}
\alias{clean_formatter}
\alias{with_formatter}
\title{Formatter orchestration}
\usage{
init_formatter(formatter)

close_formatter(formatter, info)

clean_formatter(formatter, info)

with_formatter(expr, formatter, info)
}
\arguments{
\item{formatter}{A serializer function}

\item{info}{A structure returned by \code{init_formatter()}}

\item{expr}{An expression to evaluate in the context of the formatter}
}
\value{
\code{init_formatter()} returns a opaque structure capturing information
used by the other functions. \code{close_formatter()} may return a value that
should be used as response body. \code{with_formatter()} returns the result of
\code{expr}. \code{clean_formatter()} is called for it's side effects and should only
be called if \code{close_formatter()} never evaluated.
}
\description{
These functions are for internal use and only exported to ease async
evaluation
}
\keyword{internal}
