% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber2-package.R
\docType{package}
\name{plumber2-package}
\alias{plumber2}
\alias{plumber2-package}
\title{plumber2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
