% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers.R
\name{register_serializer}
\alias{register_serializer}
\title{Register a serializer to a name for use with the \verb{@serializer} tag}
\usage{
register_serializer(name, fun, mime_type)
}
\arguments{
\item{name}{The name to register the serializer function to. If already
present the current serializer will be overwritten by the one provided by you}

\item{fun}{A function that, when called, returns a unary function that can
serialize a response body to the mime type defined in \code{mime_type}}

\item{mime_type}{The format this serializer creates. You should take care to
ensure that the value provided is a standard mime type for the format}
}
\value{
This function is called for its side effects
}
\description{
plumber2 comes with many serializers that should cover almost all standard
use cases. Still you might want to provide some of your own, which this
function facilitates.
}
\seealso{
\code{\link[=register_parser]{register_parser()}}
}
