% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers.R
\name{device_formatter}
\alias{device_formatter}
\title{Create a graphics device formatter}
\usage{
device_formatter(dev_open, dev_close = grDevices::dev.off())
}
\arguments{
\item{dev_open}{The function that opens the device}

\item{dev_close}{The function closing the device. Usually this would be
\code{\link[grDevices:dev]{grDevices::dev.off()}}}
}
\value{
A device formatter function
}
\description{
This internal function facilitates creating a formatter that uses a specific
device for rendering.
}
\keyword{internal}
