% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber2-package.R
\docType{package}
\name{plumber2-package}
\alias{plumber2}
\alias{plumber2-package}
\title{plumber2: Easy and Powerful Webservers in R}
\description{
Automatically create a webserver from annotated R files or by building it up programmatically. Provides automatic OpenAPI documentation, input handling, async support, and middleware support.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://posit-dev.github.io/plumber2/}
  \item \url{https://github.com/posit-dev/plumber2}
  \item Report bugs at \url{https://github.com/posit-dev/plumber2/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
