% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber2-package.R
\docType{package}
\name{plumber2-package}
\alias{plumber2}
\alias{plumber2-package}
\title{plumber2: Easy and Powerful Webservers}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Automatically create a webserver from annotated 'R' files or by building it up programmatically. Provides automatic 'OpenAPI' documentation, input handling, async support, and middleware support.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://plumber2.posit.co/}
  \item \url{https://github.com/posit-dev/plumber2}
  \item Report bugs at \url{https://github.com/posit-dev/plumber2/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
