% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_session_cookie.R, R/conditions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{random_key}
\alias{abort_status}
\alias{abort_http_problem}
\alias{abort_bad_request}
\alias{abort_conflict}
\alias{abort_forbidden}
\alias{abort_gone}
\alias{abort_internal_error}
\alias{abort_method_not_allowed}
\alias{abort_not_acceptable}
\alias{abort_not_found}
\alias{abort_unauthorized}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reqres}{\code{\link[reqres:abort_http_problem]{abort_bad_request}}, \code{\link[reqres:abort_http_problem]{abort_conflict}}, \code{\link[reqres:abort_http_problem]{abort_forbidden}}, \code{\link[reqres:abort_http_problem]{abort_gone}}, \code{\link[reqres]{abort_http_problem}}, \code{\link[reqres:abort_http_problem]{abort_internal_error}}, \code{\link[reqres:abort_http_problem]{abort_method_not_allowed}}, \code{\link[reqres:abort_http_problem]{abort_not_acceptable}}, \code{\link[reqres:abort_http_problem]{abort_not_found}}, \code{\link[reqres:abort_http_problem]{abort_status}}, \code{\link[reqres:abort_http_problem]{abort_unauthorized}}, \code{\link[reqres]{random_key}}}
}}

