% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{register_parser}
\alias{register_parser}
\title{Register a parser to a name for use with the \verb{@parser} tag}
\usage{
register_parser(name, fun, mime_types)
}
\arguments{
\item{name}{The name to register the parser function to. If already
present the current parser will be overwritten by the one provided by you}

\item{fun}{A function that, when called, returns a binary function that can
parse a request body. The first argument takes a raw vector with the binary
encoding of the request body, the second argument takes any additional
directives given by the requests \code{Content-Type} header}

\item{mime_types}{One or more mime types that this parser can handle. The
mime types are allowed to contain wildcards, e.g. \code{"text/*"}}
}
\value{
This function is called for its side effects
}
\description{
plumber2 comes with many parsers that should cover almost all standard
use cases. Still you might want to provide some of your own, which this
function facilitates.
}
\seealso{
\code{\link[=register_serializer]{register_serializer()}}
}
